requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.3.1
django-vite==3.0.4
djangorestframework==3.15.1
sqlparse==0.5.0
